Changes to allow using multiple client implementations simultaneously
