+++ rsfbclient-native/src/lib.rs
@@ -7,4 +7,4 @@ pub(crate) mod row;
 pub(crate) mod status;
 pub(crate) mod xsqlda;
 
-pub use connection::NativeFbClient;
+pub use connection::{Args, NativeFbClient};
