+++ src/connection/stmt_cache.rs
@@ -78,7 +78,7 @@ impl<T> StmtCache<T> {
 /// Functions specific for when the data is a `StatementData`
 impl<H> StmtCache<StatementData<H>>
 where
-    H: Clone + Copy,
+    H: Send + Clone + Copy,
 {
     /// Get a prepared statement from the cache, or prepare one
     pub fn get_or_prepare<C>(
